1.change to docker-java:1.4.0 2.fix docker dns issue
