+++ src/main/java/org/jenkinsci/plugins/dockerbuildstep/cmd/StartCommand.java
@@ -1,92 +1,51 @@
 package org.jenkinsci.plugins.dockerbuildstep.cmd;
 
+import com.github.dockerjava.api.DockerClient;
+import com.github.dockerjava.api.DockerException;
+import com.github.dockerjava.api.command.InspectContainerResponse;
 import hudson.Extension;
 import hudson.model.AbstractBuild;
 import hudson.util.FormValidation;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.Map;
-
 import org.jenkinsci.plugins.dockerbuildstep.action.DockerContainerConsoleAction;
 import org.jenkinsci.plugins.dockerbuildstep.action.EnvInvisibleAction;
 import org.jenkinsci.plugins.dockerbuildstep.log.ConsoleLogger;
 import org.jenkinsci.plugins.dockerbuildstep.util.BindParser;
-import org.jenkinsci.plugins.dockerbuildstep.util.LinkUtils;
 import org.jenkinsci.plugins.dockerbuildstep.util.PortBindingParser;
 import org.jenkinsci.plugins.dockerbuildstep.util.PortUtils;
 import org.jenkinsci.plugins.dockerbuildstep.util.Resolver;
 import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.QueryParameter;
 
-import com.github.dockerjava.api.DockerClient;
-import com.github.dockerjava.api.DockerException;
-import com.github.dockerjava.api.command.InspectContainerResponse;
-import com.github.dockerjava.api.model.Bind;
-import com.github.dockerjava.api.model.Links;
-import com.github.dockerjava.api.model.PortBinding;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Map;
 
 /**
  * This command starts one or more Docker containers. It also exports some build environment variables like IP or
  * started containers.
- * 
- * @see http://docs.docker.com/reference/api/docker_remote_api_v1.13/#start-a-container
- * 
+ *
  * @author vjuranek
- * 
+ * @see http://docs.docker.com/reference/api/docker_remote_api_v1.13/#start-a-container
  */
 public class StartCommand extends DockerCommand {
 
     private final String containerIds;
-    private final boolean publishAllPorts;
-    private final String portBindings;
     private final String waitPorts;
-    private final String links;
-    private final String bindMounts;
-    private final boolean privileged;
     private final String containerIdsLogging;
 
     @DataBoundConstructor
-    public StartCommand(String containerIds, boolean publishAllPorts, String portBindings, String waitPorts,
-            String links, String bindMounts, boolean privileged, String containerIdsLogging) {
+    public StartCommand(String containerIds, String waitPorts, String containerIdsLogging) {
         this.containerIds = containerIds;
-        this.publishAllPorts = publishAllPorts;
-        this.portBindings = portBindings;
         this.waitPorts = waitPorts;
-        this.links = links;
-        this.bindMounts = bindMounts;
-        this.privileged = privileged;
         this.containerIdsLogging = containerIdsLogging;
     }
 
     public String getContainerIds() {
         return containerIds;
     }
-
-    public boolean getPublishAllPorts() {
-        return publishAllPorts;
-    }
-
-    public String getPortBindings() {
-        return portBindings;
-    }
-
     public String getWaitPorts() {
         return waitPorts;
     }
-
-    public String getLinks() {
-        return links;
-    }
-
-    public String getBindMounts() {
-        return bindMounts;
-    }
-
-    public boolean getPrivileged() {
-        return privileged;
-    }
-
     public String getContainerIdsLogging() {
         return containerIdsLogging;
     }
@@ -99,9 +58,6 @@ public class StartCommand extends DockerCommand {
         }
 
         List<String> ids = Arrays.asList(Resolver.buildVar(build, containerIds).split(","));
-        PortBinding[] portBindingsRes = PortBindingParser.parse(Resolver.buildVar(build, portBindings));
-        Links linksRes = LinkUtils.parseLinks(Resolver.buildVar(build, links));
-        Bind[] bindsRes = BindParser.parse(Resolver.buildVar(build, bindMounts));
         List<String> logIds = Arrays.asList(Resolver.buildVar(build, containerIdsLogging).split(","));
 
         DockerClient client = getClient(build, null);
@@ -115,8 +71,7 @@ public class StartCommand extends DockerCommand {
                 outAction = attachContainerOutput(build, id);
             }
 
-            client.startContainerCmd(id).withPublishAllPorts(publishAllPorts).withPortBindings(portBindingsRes)
-                    .withLinks(linksRes.getLinks()).withBinds(bindsRes).withPrivileged(privileged).exec();
+            client.startContainerCmd(id).exec();
             console.logInfo("started container id " + id);
 
             InspectContainerResponse inspectResp = client.inspectContainerCmd(id).exec();
@@ -159,24 +114,6 @@ public class StartCommand extends DockerCommand {
         @Override
         public String getDisplayName() {
             return "Start container(s)";
-        }
-
-        public FormValidation doTestPortBindings(@QueryParameter String portBindings) {
-            try {
-                PortBindingParser.parse(portBindings);
-            } catch (IllegalArgumentException e) {
-                return FormValidation.error(e.getMessage());
-            }
-            return FormValidation.ok("OK");
-        }
-
-        public FormValidation doTestBindMounts(@QueryParameter String bindMounts) {
-            try {
-                BindParser.parse(bindMounts);
-            } catch (IllegalArgumentException e) {
-                return FormValidation.error(e.getMessage());
-            }
-            return FormValidation.ok("OK");
         }
     }
