+++ src/main/resources/org/jenkinsci/plugins/dockerbuildstep/cmd/CreateContainerCommand/config-detail.jelly
@@ -29,5 +29,23 @@
 		<f:entry field="memoryLimit" title="Memory limit" description="Memory limit (format: number{optional unit}, where unit = b, k, m or g). Example: 1024m">
 			<f:textbox />
 		</f:entry>		
+		<f:entry field="dns" title="DNS" description="Option --dns:Comma separated list of links. Example: dns1,dns2,...">
+			<f:textbox />
+		</f:entry>
+		<f:entry field="publishAllPorts" title="Publish all ports">
+			<f:checkbox />
+		</f:entry>
+		<f:entry field="portBindings" title="Port bindings">
+			<f:textarea />
+		</f:entry>
+		<f:validateButton title="${%Check syntax}" method="testPortBindings" with="portBindings" />
+		 <f:entry field="bindMounts" title="Bind mounts">
+			<f:textarea />
+		</f:entry>
+		<f:validateButton title="${%Check syntax}" method="testBindMounts" with="bindMounts" />
+		<f:entry field="privileged" title="Extended privileges">
+			<f:checkbox />
+		</f:entry>
+
 	</f:advanced>
 </j:jelly>
