+++ src/main/resources/org/jenkinsci/plugins/dockerbuildstep/cmd/CreateContainerCommand/help-privileged.html
@@ -1,0 +1,5 @@
+<div>
+  Give extended privileges to the container(s).
+  <p>
+  This setting corresponds to the <code>--privileged</code> option of the <code>docker run</code> CLI command. 
+</div>
