+++ src/main/resources/org/jenkinsci/plugins/dockerbuildstep/cmd/StartCommand/config-detail.jelly
@@ -11,30 +11,9 @@
 help file problem.
 -->
     <f:advanced align="left">
-        <f:entry field="publishAllPorts" title="Publish all ports">
-            <f:checkbox />
-        </f:entry>
 
-        <f:entry field="portBindings" title="Port bindings">
-            <f:textarea />
-        </f:entry>
-        <f:validateButton title="${%Check syntax}" method="testPortBindings" with="portBindings" />
-        
         <f:entry field="waitPorts" title="Wait for ports">
             <f:textarea />
-        </f:entry>
-        
-        <f:entry field="links" title="Links" description="Comma separated list of links. Example: name1:alias1,name1:alias2">
-            <f:textbox />
-        </f:entry>
-
-        <f:entry field="bindMounts" title="Bind mounts">
-            <f:textarea />
-        </f:entry>
-        <f:validateButton title="${%Check syntax}" method="testBindMounts" with="bindMounts" />
-
-        <f:entry field="privileged" title="Extended privileges">
-            <f:checkbox />
         </f:entry>
 
 	    <f:entry field="containerIdsLogging" title="Capture output from" description="Comma separated list of containers to attach to and capture output from.">
