--- src/main/resources/org/jenkinsci/plugins/dockerbuildstep/cmd/StartCommand/help-privileged.html
@@ -1,5 +1,0 @@
-<div>
-  Give extended privileges to the container(s).
-  <p>
-  This setting corresponds to the <code>--privileged</code> option of the <code>docker run</code> CLI command. 
-</div>
